refactor: align OpenAPI contracts and support user APIs
